[BUGFIX] When to perform `apt-get update`
Checking the last modify time of the
`/var/lib/apt/periodic/update-success-timestamp` file for when the last
`apt-get` update was performed is inaccurate. Hence, we switch to
checking the last modify time of the `/var/cache/apt/pkgcache.bin` file,
suggested by this answer:

    http://askubuntu.com/a/410259

Instead of performing the check AND the `apt-get update` in the
`mylib_perform_apt_update` module, we use the module to perform the
check and do the `apt-get update` using the `apt` module.

In light of this split of duties, we rename the
`mylib_perform_apt_update` module to `check_if_need_to_apt_get_update.py`

This line is required in `check_if_need_to_apt_get_update.py`:

    # vim: set fileencoding=utf-8

probably because we pasted the outout of a
`stat /var/cache/apt/pkgcache.bin` command, which contains some UTF-8
character.

Signed-off-by: Yan Han Pang <539739+yanhan@users.noreply.github.com>